fix(resilience): retire standalone interval seeder
